Move static graph tests into qa_Block
